# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API(Laravel 12)",
        "description": "This project is a simple RESTful API built with Laravel 12, focused on user authentication and checklist management. It provides features such as user registration and login, along with the ability to create and manage checklists and their items.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/checklist/{list_id}/item": {
            "post": {
                "tags": [
                    "Todo(ChecklistItem)"
                ],
                "summary": "Add new item to checklist",
                "description": "Create a new checklist item under the specified checklist. Requires Bearer token.",
                "operationId": "addChecklistItem",
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "Checklist ID to which the item will be added",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Buy milk Description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Checklist not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/checklist": {
            "get": {
                "tags": [
                    "Todo(Checklist)"
                ],
                "summary": "Get all checklists",
                "description": "Returns a paginated list of checklists. Requires Bearer token.",
                "operationId": "getChecklists",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by checklist name or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Checklist"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Todo(Checklist)"
                ],
                "summary": "Create a new checklist",
                "description": "Create a new checklist with name and description. Requires Bearer token.",
                "operationId": "createChecklist",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Daily Tasks"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Tasks for daily routine"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Checklist created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/checklist/{id}/item": {
            "get": {
                "tags": [
                    "Todo(ChecklistItem)"
                ],
                "summary": "Get all checklist item",
                "description": "Retrieve detailed information of a checklist by its ID. Requires Bearer token.",
                "operationId": "getChecklistDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the checklist",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist detail retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/checklist/{id}": {
            "put": {
                "tags": [
                    "Todo(Checklist)"
                ],
                "summary": "Update checklist",
                "description": "Update an existing checklist by ID. Requires Bearer token.",
                "operationId": "updateChecklist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated checklist title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated checklist description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checklist updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChecklistItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Todo(Checklist)"
                ],
                "summary": "Delete checklist",
                "description": "Delete a checklist by its ID. Requires Bearer token.",
                "operationId": "deleteChecklist",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist ID to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Todo Deleted!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ChecklistItem": {
                "title": "Checklist Item",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Buy groceries"
                    },
                    "description": {
                        "type": "string",
                        "example": "Things to buy at the market"
                    },
                    "status": {
                        "type": "string",
                        "example": "ONGOING"
                    }
                },
                "type": "object"
            },
            "Checklist": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Daily Tasks"
                    },
                    "description": {
                        "type": "string",
                        "example": "Tasks to complete daily"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChecklistItem"
                        }
                    },
                    "owner": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}